+++ src/Logging/Logging.Debug/DebugLogService.cs
@@ -2,12 +2,9 @@
 {
     internal sealed class DebugLogService : ScopedLogService<DebugLogOptions>
     {
-        private readonly LogMessageFormatter _formatter;
-
         internal DebugLogService(string context)
             : base(context)
         {
-            _formatter = new LogMessageFormatter();
         }
 
         protected override void PublishMessage(LogMessage message)
@@ -15,7 +12,7 @@
             try
             {
                 System.Diagnostics.Debug.WriteLine(
-                        _formatter.Format(
+                        MessageFormatter.Format(
                             message,
                             Options.TimeStampFormat));
             }
